# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Library Management API",
        "description": "Library Management API Documentation"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/books": {
            "get": {
                "summary": "Retrieve a list of books",
                "responses": {
                    "200": {
                        "description": "A list of books"
                    }
                }
            }
        },
        "/books/borrow": {
            "post": {
                "summary": "Borrow a book",
                "parameters": [
                    {
                        "in": "body",
                        "name": "borrow",
                        "description": "Borrow a book",
                        "schema": {
                            "type": "object",
                            "required": [
                                "memberCode",
                                "bookCode"
                            ],
                            "properties": {
                                "memberCode": {
                                    "type": "string"
                                },
                                "bookCode": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book borrowed successfully"
                    }
                }
            }
        },
        "/books/return": {
            "post": {
                "summary": "Return a book",
                "parameters": [
                    {
                        "in": "body",
                        "name": "return",
                        "description": "Return a book",
                        "schema": {
                            "type": "object",
                            "required": [
                                "memberCode",
                                "bookCode",
                                "returnDate"
                            ],
                            "properties": {
                                "memberCode": {
                                    "type": "string"
                                },
                                "bookCode": {
                                    "type": "string"
                                },
                                "returnDate": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book returned successfully"
                    }
                }
            }
        },
        "/members/{code}/penalties": {
            "get": {
                "summary": "Get penalties for a member",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Member code"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Penalties retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/members": {
            "get": {
                "summary": "Retrieve a list of members",
                "responses": {
                    "200": {
                        "description": "A list of members"
                    }
                }
            }
        },
        "/members/{code}": {
            "get": {
                "summary": "Retrieve a member's details",
                "parameters": [
                    {
                        "in": "path",
                        "name": "code",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Member code"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member details"
                    }
                }
            }
        }
    }
}